Separate RBD_Light object construction from the pinMode() setup function.
The pinMode() setup should happen in the global setup() function, but
the constructor of the RBD_Light object is called earlier upon global
object initialization. Hence, setting up pins in the constructor is too
early. Instead, the pins should be left alone and set up by a second function
call later. For code backward compatibility, the old usage of the constructor
is still allowed, though.